+++ src/commands/CopyCommand.ts
@@ -1,0 +1,20 @@
+import * as clipboardy from "clipboardy";
+import { SolutionExplorerProvider } from "../SolutionExplorerProvider";
+import { TreeItem, ContextValues } from "../tree";
+import { CommandBase } from "./base/CommandBase";
+
+export class CopyCommand extends CommandBase {
+    
+    constructor(private readonly provider: SolutionExplorerProvider) {
+        super();
+    }
+
+    protected shouldRun(item: TreeItem): boolean {
+        if (item && item.path) return true;
+        return false;
+    }
+
+    protected async runCommand(item: TreeItem, args: string[]): Promise<void> {    
+        await clipboardy.write(item.path);
+    }
+}