+++ src/model/Projects/Project.ts
@@ -32,7 +32,7 @@ export abstract class Project {
 
     public abstract deleteFile(filepath: string): Promise<void>;
 
-    public abstract createFile(folderpath: string, filename: string): Promise<string>;
+    public abstract createFile(folderpath: string, filename: string, content?: string): Promise<string>;
 
     public abstract renameFolder(folderpath: string, name: string): Promise<string>;